revert page state directory refactor
